Added Challenged 24 and 25

+++ Enumerators/Fibbonacci.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 using ProjectEuler;
@@ -10,14 +11,14 @@ namespace Enumerators
 {
     public static class Fibonacci
     {
-        public static IEnumerable<long> Sequence()
+        public static IEnumerable<BigInteger> Sequence()
         {
-            long primaryTerm = 0;
-            long secondaryTerm = 1;
+            BigInteger primaryTerm = 0;
+            BigInteger secondaryTerm = 1;
            
             do
             {
-                long oldNumber2 = secondaryTerm;
+                BigInteger oldNumber2 = secondaryTerm;
                 secondaryTerm = primaryTerm + secondaryTerm;
                 yield return secondaryTerm;
                 primaryTerm = oldNumber2;
