+++ UnitTests/EnumeratorTests/FibbonacciTest.cs
@@ -12,14 +12,14 @@ namespace EnumeratorTests
         [Test]
         public void Test10thFibbonacci()
         {
-            long result = Fibonacci.Sequence().Take(9).Last();
+            long result = (long)Fibonacci.Sequence().Take(9).Last();
             Assert.AreEqual(55, result);
         }
 
         [Test]
         public void Test20thFibbonacci()
         {
-            long result = Fibonacci.Sequence().Take(19).Last();
+            long result = (long)Fibonacci.Sequence().Take(19).Last();
             Assert.AreEqual(6765, result);
         }
     }
